backend/requirements.txt
anyio==4.10.0
click==8.3.0
fastapi==0.95.2
h11==0.16.0
idna==3.10
pydantic==1.10.13
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.36.0
